backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and storage
supabase==2.3.0
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# External integrations
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Audio processing for voice interface
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# Background tasks
celery==5.3.4
celery[redis]==5.3.4

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
qrcode[pil]==7.4.2
cryptography==41.0.8

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring and logging
structlog==23.2.0